[No-issue] Fix responsiveness for multi-viewport (#12)
* fix(mobile): resize book fonts and spacing for better mobile visibility

Read toggle padding was too close to delete book button. Added a thicker padding to decrease error
clicks.

* feat(book_card): add confirmation popup when deleting a book

This better prevents a user from removing their hard, entered book and not making it a pain to re-do
their progress.

* style(mobile): set up responsiveness for book addition form

Works with column grid breakpoints to offer a smooth experience on mobile and desktop as far as
usability goes by providing pleasant sizes to work with based on the viewport width.

* style(*): add icons for a bit of fancy styling

Icons added to the main functionality buttons. * Add a book, * Close form, * Add book to collection,
* Heading.

+++ js/content/bookCard.js
@@ -58,7 +58,7 @@ function buildBookCard(book) {
   );
 
   const readToggleRow = buildDiv();
-  readToggleRow.classList.add('row', 'pt-2', 'pb-1', 'readToggleRow');
+  readToggleRow.classList.add('row', 'pt-2', 'pb-2', 'readToggleRow');
 
   const readColumn = buildDiv();
   readColumn.classList.add('col', 'h4', 'd-flex', 'align-items-center', 'justify-content-center', 'mb-0');
@@ -103,9 +103,11 @@ function buildBookCard(book) {
   trashIcon.classList.add('bi', 'bi-trash');
   deleteButton.append(trashIcon);
   deleteButton.addEventListener('click', () => {
-    deleteBook(book)
-    reRenderBooks();
-    deleteButton.removeEventListener('click');
+    if (window.confirm("Are you sure you'd like to remove this book from your collection?")) {
+      deleteBook(book)
+      reRenderBooks();
+      deleteButton.removeEventListener('click');
+    }
   })
 
   buttonRow.append(deleteButton)
